requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.1
joblib==1.4.2
mlflow==2.16.2
prometheus-client==0.21.0
pydantic==2.9.2
pytest==8.3.3
httpx==0.27.2
python-dotenv==1.0.1